fix(solana_v2): add missing exports
